Socket.IO + EmailConfirmation token + MailService

+++ src/main/java/www/aaltogetherbackend/controllers/AuthController.java
@@ -7,10 +7,8 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+import www.aaltogetherbackend.models.EmailConfirmationToken;
 import www.aaltogetherbackend.models.RefreshToken;
 import www.aaltogetherbackend.models.User;
 import www.aaltogetherbackend.payloads.requests.LoginRequest;
@@ -20,8 +18,7 @@ import www.aaltogetherbackend.payloads.responses.ErrorMessageResponse;
 import www.aaltogetherbackend.payloads.responses.LoginResponse;
 import www.aaltogetherbackend.payloads.responses.MessageResponse;
 import www.aaltogetherbackend.repositories.UserRepository;
-import www.aaltogetherbackend.services.JwtUtils;
-import www.aaltogetherbackend.services.RefreshTokenService;
+import www.aaltogetherbackend.services.*;
 
 import java.util.Optional;
 
@@ -34,13 +31,17 @@ public class AuthController {
     private final AuthenticationManager authenticationManager;
     private final PasswordEncoder encoder;
     private final RefreshTokenService refreshTokenService;
+    private final MailService mailService;
+    private final EmailConfirmationTokenService emailConfirmationTokenService;
 
-    public AuthController(UserRepository userRepository, JwtUtils jwtUtils, AuthenticationManager authenticationManager, PasswordEncoder encoder, RefreshTokenService refreshTokenService) {
+    public AuthController(UserRepository userRepository, JwtUtils jwtUtils, AuthenticationManager authenticationManager, PasswordEncoder encoder, RefreshTokenService refreshTokenService, MailService mailService, EmailConfirmationTokenService emailConfirmationTokenService) {
         this.userRepository = userRepository;
         this.jwtUtils = jwtUtils;
         this.authenticationManager = authenticationManager;
         this.encoder = encoder;
         this.refreshTokenService = refreshTokenService;
+        this.mailService = mailService;
+        this.emailConfirmationTokenService = emailConfirmationTokenService;
     }
 
     @PostMapping("/login")
@@ -52,6 +53,11 @@ public class AuthController {
         String jwt = jwtUtils.generateToken(loginRequest.username());
 
         User user = (User) authentication.getPrincipal();
+
+        if (!user.isEmailVerified()) {
+            return ResponseEntity.badRequest().body(new ErrorMessageResponse("Email is not verified!"));
+        }
+
         refreshTokenService.deleteByUser(user);
         String refreshToken = refreshTokenService.generateRefreshToken(loginRequest.username());
         return ResponseEntity.ok().body(new LoginResponse("You've been signed in!", jwt, refreshToken));
@@ -69,7 +75,12 @@ public class AuthController {
         user.setUsername(signupRequest.username());
         user.setPassword(encoder.encode(signupRequest.password()));
         user.setEmail(signupRequest.email());
+
         userRepository.save(user);
+
+        String token = emailConfirmationTokenService.generateEmailVerificationToken(user);
+
+        mailService.SendMail("test@example.com", "Email Verification", "Click here to verify your email: http://localhost:8080/api/auth/verify-email?token=" + token);
         return ResponseEntity.ok().body(new MessageResponse("User registered successfully!"));
     }
 
@@ -89,7 +100,7 @@ public class AuthController {
         return ResponseEntity.ok().body(new LoginResponse("Token refreshed!", jwt, refreshTokenRequest.refreshToken()));
     }
 
-    @PostMapping("/logout")
+    @GetMapping("/logout")
     public ResponseEntity<?> logout() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         User user = (User) authentication.getPrincipal();
@@ -99,4 +110,43 @@ public class AuthController {
         return ResponseEntity.ok().body(new MessageResponse("You've been signed out!"));
     }
 
+    @GetMapping("/verify-email")
+    public ResponseEntity<?> verifyEmail(@Valid @RequestParam String token) {
+        Optional<EmailConfirmationToken> emailConfirmationToken = emailConfirmationTokenService.findByToken(token);
+
+        if (emailConfirmationToken.isEmpty()) {
+            return ResponseEntity.badRequest().body(new ErrorMessageResponse("Token is invalid!"));
+        }
+
+        if (emailConfirmationTokenService.isExpired(token)) {
+            return ResponseEntity.badRequest().body(new ErrorMessageResponse("Token is expired!"));
+        }
+
+        User user = emailConfirmationToken.get().getUser();
+        user.setEmailVerified(true);
+        userRepository.save(user);
+        return ResponseEntity.ok().body(new MessageResponse("Email verified!"));
+    }
+
+    /*
+    @PostMapping("/forgot-password")
+    public ResponseEntity<?> forgotPassword(@Valid @RequestBody String email) {
+        User user = userRepository.findByEmail(email);
+        if (user == null) {
+            return ResponseEntity.badRequest().body(new ErrorMessageResponse("User not found!"));
+        }
+        String token = UUID.randomUUID().toString();
+        userRepository.save(user);
+        mailService.SendMail("test@example.com", "Password Reset", "Click here to reset your password: http://localhost:8080/api/auth/reset-password?token=" + token);
+        return ResponseEntity.ok().body(new MessageResponse("Password reset email sent!"));
+    }
+
+    @PostMapping("/reset-password")
+    public ResponseEntity<?> resetPassword(@Valid @RequestParam String token, @Valid @RequestBody String password) {
+        User user = userRepository.findByEmailVerificationToken(token);
+        user.setPassword(encoder.encode(password));
+        userRepository.save(user);
+        return ResponseEntity.ok().body(new MessageResponse("Password reset successfully!"));
+    }
+    */
 }
